feat: introduce reviewer_service_helpers for improved code organization
- Added a new module `reviewer_service_helpers.py` to encapsulate internal helper functions for the reviewer agent loop, enhancing the readability and maintainability of the `reviewer_service.py`.
- Functions include JSON parsing, token management, and per-round LLM calls, streamlining the review process and separating concerns within the codebase.
- This refactor allows the main orchestrator function `run_review_loop` to remain focused on its primary responsibilities, improving overall code clarity.

This change contributes to a more structured and efficient implementation of the reviewer agent functionality.
